Merge up to r13-7553-g1e6a948cd22f2f142cdc828296f78c7af9e283c8 (10th July 2023)

+++ libstdc++-v3/testsuite/27_io/basic_istream/extractors_arithmetic/char/09.cc
@@ -37,7 +37,7 @@ void test09()
   (is>>std::ws) >> c;
   (is>>std::ws) >> f2;
   VERIFY( f1 == 2456 );
-  VERIFY( f2 == 0.00567 );
+  VERIFY( f2 == (double) 0.00567 ); // N.B. cast removes excess precision
   VERIFY( c == '-' );
 }
